// src/app/modules/comercial/materiais-perdidos/models/material-perdido.ts
export interface MaterialPerdido {
  codAssociacao: number;
  codConcorrente: number;
  nomeConcorrente: string;
  codMaterialManetoni: number;
  desMaterialManetoni: string;
  vrConcorrente: number;
  vrManetoni: number;
  percentualDiferenca: number;
  nrProposta: number;
  codEmpresa: number;
  nomeEmpresa: string;
  codUsuarioCadastro: number;
  nomeUsuarioCadastro: string;
  dataCadastro: string;
}